[4.21] Storage: replace repetitive code with get_dv_size_from_datasource fuction (#4215)
##### Short description:
Manual cherry-pick:
https://github.com/RedHatQE/openshift-virtualization-tests/pull/4153

##### More details:

https://redhat.atlassian.net/browse/CNV-82173

##### What this PR does / why we need it:

##### Which issue(s) this PR fixes:

##### Special notes for reviewer:

##### jira-ticket:
<!-- full-ticket-url needs to be provided. This would add a link to the
pull request to the jira and close it when the pull request is merged
If the task is not tracked by a Jira ticket, just write "NONE".
-->
