+++ tests/storage/utils.py
@@ -10,7 +10,6 @@ from ocp_resources.cdi import CDI
 from ocp_resources.cluster_role import ClusterRole
 from ocp_resources.config_map import ConfigMap
 from ocp_resources.daemonset import DaemonSet
-from ocp_resources.data_source import DataSource
 from ocp_resources.datavolume import DataVolume
 from ocp_resources.hostpath_provisioner import HostPathProvisioner
 from ocp_resources.pod import Pod
@@ -477,12 +476,3 @@ def create_windows_directory(windows_vm: VirtualMachineForTests, directory_path:
         windows_vm=windows_vm,
         directory_path=directory_path,
     )
-
-
-def get_dv_size_from_datasource(data_source: DataSource) -> str | int | None:
-    source_dict = data_source.source.instance.to_dict()
-    source_spec_dict = source_dict["spec"]
-    dv_size = source_spec_dict.get("resources", {}).get("requests", {}).get("storage") or source_dict.get(
-        "status", {}
-    ).get("restoreSize")
-    return dv_size
